Fix bugs

+++ Office events-final project/src/main/java/com/fmi/spring5/service/UserServiceImpl.java
@@ -3,19 +3,15 @@ package com.fmi.spring5.service;
 import com.fmi.spring5.exceptions.EntityAlreadyExistsException;
 import com.fmi.spring5.exceptions.InvalidArgumentException;
 import com.fmi.spring5.exceptions.NoSuchEntityException;
-import com.fmi.spring5.model.Admin;
-import com.fmi.spring5.model.Dev;
-import com.fmi.spring5.model.Manager;
-import com.fmi.spring5.model.User;
-import com.fmi.spring5.repository.AdminRepository;
-import com.fmi.spring5.repository.DevRepository;
-import com.fmi.spring5.repository.ManagerRepository;
-import com.fmi.spring5.repository.UserRepository;
+import com.fmi.spring5.model.*;
+import com.fmi.spring5.repository.*;
 import com.fmi.spring5.utils.Role;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -28,13 +24,19 @@ public class UserServiceImpl implements UserService {
     private DevRepository devRepository;
     @Autowired
     private AdminRepository adminRepository;
+    @Autowired
+    private TeamRepository teamRepository;
 
     @Override
     public Dev addUser(User user) throws EntityAlreadyExistsException {
-        userRepository.findUserByUsername(user.getUsername())
-                .orElseThrow(() -> new EntityAlreadyExistsException(String.format("The user with name: '%s' already exist", user.getUsername())));
 
-        return devRepository.save(new Dev(user));
+        Optional<User> foundUser = userRepository.findUserByUsername(user.getUsername());
+
+        if (foundUser.isPresent()) {
+            throw new EntityAlreadyExistsException(String.format("The user with name: '%s' already exist", user.getUsername()));
+
+        }
+        return devRepository.save(new Dev(user, true));
     }
 
     @Override
@@ -44,6 +46,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User getUser(String username) throws NoSuchEntityException {
+
         return userRepository.findUserByUsername(username).
                 orElseThrow(() -> new NoSuchEntityException(String.format("The user with name: '%s' doesn't exist", username)));
     }
@@ -72,38 +75,47 @@ public class UserServiceImpl implements UserService {
         if (userRole.equals(Role.ADMIN.getStringRole())) {
             if (newRole.equals(Role.MANAGER.getStringRole())) {
                 adminRepository.deleteAdminByUser(foundUser);
-                managerRepository.save(new Manager(foundUser));
+                managerRepository.save(new Manager(foundUser, false));
             } else if (newRole.equals(Role.DEV.getStringRole())) {
                 adminRepository.deleteAdminByUser(foundUser);
-                devRepository.save(new Dev(foundUser));
+                devRepository.save(new Dev(foundUser, false));
             }
         } else if (userRole.equals(Role.MANAGER.getStringRole())) {
+            List<Team> teams =teamRepository.findByManager(managerRepository.findByUser(foundUser).get());
+            if(!teams.isEmpty()){
+                throw new InvalidArgumentException("The manager has assigned teams. Can not change the role!");
+            }
             if (newRole.equals(Role.ADMIN.getStringRole())) {
                 managerRepository.deleteByUser(foundUser);
-                adminRepository.save(new Admin(foundUser));
+                adminRepository.save(new Admin(foundUser, false));
             } else if (newRole.equals(Role.DEV.getStringRole())) {
                 managerRepository.deleteByUser(foundUser);
-                devRepository.save(new Dev(foundUser));
+                devRepository.save(new Dev(foundUser, false));
             }
         } else {
             if (newRole.equals(Role.ADMIN.getStringRole())) {
                 devRepository.deleteByUser(foundUser);
-                adminRepository.save(new Admin(foundUser));
+                adminRepository.save(new Admin(foundUser, false));
             } else if (newRole.equals(Role.MANAGER.getStringRole())) {
                 devRepository.deleteByUser(foundUser);
-                managerRepository.save(new Manager(foundUser));
+                managerRepository.save(new Manager(foundUser, false));
             }
         }
     }
 
     @Override
     public User update(User user) throws NoSuchEntityException {
-        User foundUser = userRepository.findById(user.getId())
+        User foundUser = userRepository.findUserByUsername(user.getUsername())
                 .orElseThrow(() -> new NoSuchEntityException(String.format("User with ID='%s' does not exist.", user.getId())));
 
         user.setId(foundUser.getId());
         user.setRole(foundUser.getRole());
         return userRepository.save(user);
+    }
+
+    @Override
+    public Iterable<Dev> getAllDev() {
+        return devRepository.findAll();
     }
 
 }
